chore: move abstract class to its own file
